* master: (26 commits)
  Add documentation
  Only delete avatar if the user has one
  Fix typo. Closes #758
  Enable DebugToolBar from settings file
  Instal DebugToolbar
  Fix CSRF for reset key
  Use new CSRF protection
  Fix library
  Refactor
  Fixes for the module
  Use new CSRF token for uploader customized forms.
  Install dependencies when testing
  Remove unnecessary libraries
  Use new Flask-WTF to support CSRF protection in customized forms
  Use new Flask-WTF
  Fix typo
  ignore
  Include a default rank for users that have not contributed
  Remove comments
  Ignore sql files
  ...

Conflicts:
	pybossa/view/account.py

+++ pybossa/view/account.py
@@ -36,8 +36,10 @@ from flask import render_template, current_app
 from flask.ext.login import login_required, login_user, logout_user, \
     current_user
 from flask.ext.mail import Message
-from flaskext.wtf import Form, TextField, PasswordField, validators, \
-    IntegerField, HiddenInput, SelectField, BooleanField, FileField
+from flask_wtf import Form
+from wtforms import TextField, PasswordField, validators, \
+    IntegerField, SelectField, BooleanField, FileField
+from wtforms.widgets import HiddenInput
 
 import pybossa.validator as pb_validator
 import pybossa.model as model
@@ -478,6 +480,7 @@ def update_profile():
             file = request.files['avatar']
             coordinates = (avatar_form.x1.data, avatar_form.y1.data,
                            avatar_form.x2.data, avatar_form.y2.data)
+
             prefix = time.time()
             file.filename = "%s_avatar.png" % prefix
             container = "user_%s" % current_user.id
@@ -485,7 +488,8 @@ def update_profile():
                                  container=container,
                                  coordinates=coordinates)
             # Delete previous avatar from storage
-            uploader.delete_file(current_user.info['avatar'], container)
+            if current_user.info.get('avatar'):
+                uploader.delete_file(current_user.info['avatar'], container)
             current_user.info = {'avatar': file.filename,
                                  'container': container}
             db.session.commit()
